requirements.txt
Django==5.0
typing_extensions==4.12.2
sqlparse==0.5.1
djangorestframework==3.15.2
selenium==4.24.0
beautifulsoup4==4.12.3